ed25519: use namespaced features for `serde_bytes`; MSRV 1.60 (#628)
Since workspaces have to be bumped to MSRV 1.60 at the same time, this
also bumps the MSRV of other crates in this repo along with bumping
their version numbers to `-pre`.

+++ dsa/src/verifying_key.rs
@@ -59,7 +59,7 @@ impl VerifyingKey {
 
         let w = s.mod_inverse(q)?.to_biguint().unwrap();
 
-        let n = (q.bits() / 8) as usize;
+        let n = q.bits() / 8;
         let block_size = hash.len(); // Hash function output size
 
         let z_len = min(n, block_size);
